修复各种BUG
以前出现扇区不正确，是没有成功的使用LBA模式，我竟然没发现还是CHS模式，而且竟然还成功工作了。第二点就是另外一个读取是有问题的，目前还没找出来具体在哪，新装的bochs还不知道怎么开调试模式。然后就是build.bat给更新好了，本来准备用cmake的，暂时还是算了，怎么方便怎么来

+++ kernel/type.h
@@ -1,5 +1,5 @@
 #define NULL (void*)0;
-typedef unsigned long long size_t;
+typedef unsigned long size_t;
 typedef unsigned int u32;
 typedef unsigned short u16;
 typedef unsigned char u8;
